Correct issues with Dashboard-List items, create, and Left-Nav
* Add Notebooks to list-sources
* Disable (commented out) creators for Panels, Notebooks (will be re-enabled in those PRs)

* Remove Observability Left-Nav group (will be re-instated in subsequent release)
